requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.2
pandas==2.1.3
scipy==1.11.4
gunicorn==21.2.0

